case ccro dms codes

+++ application/views/users/add.php
@@ -66,7 +66,7 @@ else{
 if(!$record['id']){
 	?>
 	<tr>
-	<td class='font18 bold'>Add a New Record</td>
+	<td class='font18 bold'>Add a New User</td>
 	<td></td>
 	</tr>
 	<?php
@@ -74,7 +74,7 @@ if(!$record['id']){
 else{
 	?>
 	<tr>
-	<td class='font18 bold'>Edit Record</td>
+	<td class='font18 bold'>Edit User</td>
 	<td></td>
 	</tr>
 	<?php
@@ -90,8 +90,18 @@ else{
 		  <td><input type="text" name="name" size="40"></td>
 		</tr>
 		-->
-		<tr class="even required"><td>* Email Login:</td><td><input type="text" name="email" size="40"></td></tr>
+		<tr class="even required"><td>Name:</td><td><input type="text" name="name" size="40"></td></tr>
 		<?php
+		if(trim($record['id'])){
+			?>
+			<tr class="even required"><td>* Login Name:</td><td><?php echo $record['email']; ?></td></tr>
+			<?php
+		}
+		else{
+			?>
+				<tr class="even required"><td>* Login Name:</td><td><input type="text" name="email" size="40"></td></tr>
+			<?php
+		}
 		if($record['id']){
 			?>
 			<tr class="even required"><td>* Password:</td><td><input type="text" name="password" size="40" placeholder='Enter new password to change current password'></td></tr>
@@ -102,11 +112,39 @@ else{
 			<tr class="even required"><td>* Password:</td><td><input type="text" name="password" size="40"></td></tr>
 			<?php
 		}
-		?>
+		?>	
 	</table>
 </td>
 <td width='50%'>
-	&nbsp;
+	<table width="100%">
+	<tr class="even required"><td style="width:110px;">User Groups:</td><td>
+		<?php
+		$t = count($user_groups);
+		$can_update_usergroup = $this->user_validation->validate("users", "setUserGroups", false);
+		//echo $can_update_usergroup;
+		for($i=0; $i<$t; $i++){
+			if(in_array($user_groups[$i]['user_group'], $user_user_groups)){
+				if($can_update_usergroup){
+					?><input checked name="user_groups[]" type="checkbox" value="<?php echo htmlentitiesX($user_groups[$i]['user_group']); ?>" ><?php echo htmlentitiesX($user_groups[$i]['user_group']); ?><br /><?php
+				}
+				else{
+					?><input style='display:none' checked name="user_groups[]" type="checkbox" value="<?php echo htmlentitiesX($user_groups[$i]['user_group']); ?>" ><?php echo htmlentitiesX($user_groups[$i]['user_group']); ?><br /><?php
+				}
+			}
+			else if($can_update_usergroup){
+				?><input name="user_groups[]" type="checkbox" value="<?php echo htmlentitiesX($user_groups[$i]['user_group']); ?>" ><?php echo htmlentitiesX($user_groups[$i]['user_group']); ?><br /><?php
+			}
+		}
+		?>
+	</td></tr>
+	<?php
+	if($record['id']){
+		?>
+		<tr class="even required"><td style="width:90px;">Date/Time Added:</td><td><?php echo date("M d, Y H:i:s", strtotime($record['dateadded']) ); ?></td></tr>
+		<?php
+	}
+	?>
+	</table>
 </td>
 </tr>
 
@@ -118,7 +156,7 @@ else{
 				<input type="button" id='savebutton' value="Save" onclick="saveRecord()" />
 			</td>
 			<?php 
-			if($record['id']!=$_SESSION['user']['id']&&$record['email']!="admin"){ //should not be able to delete self and admin
+			if($record['id']&&$record['id']!=$_SESSION['user']['id']&&$record['email']!="admin"){ //should not be able to delete self and admin
 				?><td><input type="button" style='background:red; color:white' value="Delete" onclick="deleteRecord('<?php echo $record['id']; ?>')" /></td><?php
 			}
 			?>
@@ -141,6 +179,8 @@ if(is_array($pictures)){
 if($record){
 	foreach($record as $key=>$value){	
 		if($key=='password'){
+		}
+		else if($key=='email'){
 		}
 		else if(trim($value)||1){
 			?>
